requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
akshare==1.9.99
alabaster==0.7.13
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.6.3
arrow==1.2.3
ast-comments==1.0.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs==22.2.0
Authlib==0.15.6
Babel==2.12.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
baostock==0.8.8
beautifulsoup4==4.12.2
binance==0.3
black @ file:///croot/black_1680737249031/work
blosc==1.11.1
blosc2==2.0.0
bokeh==3.1.1
browser-cookie3==0.18.2
bs4==0.0.1
cachetools==4.2.2
cattrs==22.2.0
ccxt==3.0.75
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi==1.15.1
cfscrape==2.1.1
charset-normalizer==3.1.0
click @ file:///tmp/build/80754af9/click_1646056706450/work
colorama==0.4.6
comm @ file:///croot/comm_1671231121260/work
contourpy==1.0.7
coverage==7.2.2
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.34
dataclasses-json==0.5.7
dateparser==1.1.8
debugpy @ file:///home/builder/ci_310/debugpy_1640789504635/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
discord==2.2.2
discord.py==2.2.2
dnspython==2.3.0
docutils==0.18.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.95.1
fastjsonschema==2.16.3
finta==1.3
fonttools==4.39.3
-e git+https://github.com/freqtrade/freqtrade.git@8364fc1bd27f0e38d110c83dab0f9142d126a510#egg=freqtrade
frozendict==2.3.7
frozenlist==1.3.3
gensim==4.3.1
gitdb==4.0.10
GitPython==0.3.6
google-auth==2.17.3
google-auth-oauthlib==1.0.0
greenlet==2.0.2
gspread==5.8.0
gspread-dataframe==3.3.0
gspread-formatting==1.1.2
h11==0.14.0
html5lib==1.1
idna==3.4
imagesize==1.4.1
imgkit==1.2.3
iniconfig==2.0.0
install==1.3.5
ipykernel @ file:///croot/ipykernel_1671488378391/work
ipython @ file:///croot/ipython_1676582224036/work
janus==1.0.0
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonpath==0.82
jsonschema==4.17.3
jupyter_client @ file:///croot/jupyter_client_1680171862562/work
jupyter_core @ file:///croot/jupyter_core_1679906564508/work
kaleido==0.2.1
kiwisolver==1.4.4
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mplfinance==0.12.9b7
msgpack==1.0.5
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nbformat==5.8.0
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
newtulipy==0.4.6
nltk==3.8.1
nose==1.3.7
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.7
openpyxl==3.1.2
orjson==3.8.10
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
pandas-ta==0.3.14b0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec @ file:///croot/pathspec_1674681560568/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
platformdirs @ file:///opt/conda/conda-bld/platformdirs_1662711380096/work
plotly==5.14.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
py-find-1st==1.1.5
py-mini-racer==0.6.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
pycoingecko==3.1.0
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.18.0
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.3
pyparsing==3.0.9
pypi-publisher==0.0.4
pypinyin==0.48.0
pyrsistent==0.19.3
pysnowball==0.1.2
pytest==7.2.2
pytest-asyncio==0.21.0
python-binance==1.0.17
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
python-rapidjson==1.10
python-telegram-bot==13.15
python-twitter-v2==0.8.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
questionary==1.10.0
redis==4.5.4
regex==2023.5.5
requests==2.28.2
requests-cache==1.0.1
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
sdnotify==0.3.2
seaborn==0.12.2
simplejson==3.19.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.10
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.26.1
ta==0.10.2
TA-Lib==0.4.26
tables==3.8.0
tabulate==0.9.0
technical==1.4.0
tenacity==8.2.2
threadpoolctl==3.1.0
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
tornado==6.1
tqdm==4.65.0
tradingpattern==0.0.1
tradingview-ta==3.3.0
traitlets @ file:///croot/traitlets_1671143879854/work
tushare==1.2.89
tweepy==4.13.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
url-normalize==1.4.3
urllib3==1.26.15
uvicorn==0.21.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==0.57.0
websockets==11.0.2
xlrd==2.0.1
xyzservices==2023.2.0
yarl==1.8.2
yfinance==0.2.18
